tests/test_tester.py
```python
import unittest
from src import tester

class TestTester(unittest.TestCase):

    def setUp(self):
        self.tester = tester.Tester()

    def test_scrape_data(self):
        result = self.tester.scrape_data()
        self.assertIsNotNone(result)

    def test_extract_data(self):
        result = self.tester.extract_data()
        self.assertIsNotNone(result)

    def test_clean_data(self):
        result = self.tester.clean_data()
        self.assertIsNotNone(result)

    def test_store_data(self):
        result = self.tester.store_data()
        self.assertIsNotNone(result)

    def test_handle_errors(self):
        result = self.tester.handle_errors()
        self.assertIsNotNone(result)

    def test_log_actions(self):
        result = self.tester.log_actions()
        self.assertIsNotNone(result)

    def test_schedule_bot(self):
        result = self.tester.schedule_bot()
        self.assertIsNotNone(result)

    def test_update_bot(self):
        result = self.tester.update_bot()
        self.assertIsNotNone(result)

    def test_documentation(self):
        result = self.tester.documentation()
        self.assertIsNotNone(result)

    def test_deploy_bot(self):
        result = self.tester.deploy_bot()
        self.assertIsNotNone(result)

if __name__ == '__main__':
    unittest.main()
```